Improves file scanning and quarantine process
Refactors the file scanning process to use sha256sum and signatures.db for malware detection.

Implements file quarantine functionality using the `aislar` utility.

Updates API endpoints for file scanning and quarantine to return informative messages.

Fixes errors in file path handling and memory management.

Improves the process of restoring and deleting quarantined files.

The `mem_percentage` field in the `proc` struct is updated to `unsigned int`.

+++ Fase3/daemon/syscalls_usac.h
@@ -6,7 +6,7 @@
 struct proc {
     char name[16];                // Process name
     pid_t pid;                    // Process ID
-    unsigned long mem_percentage; // Memory usage percentage
+    unsigned int mem_percentage; // Memory usage percentage
 };
 
 struct page_faults {
